demo_app/requirements.txt
Pillow
matplotlib
pandas
numpy
